Move experimental Downloader implementation into the http worker
The big Downloader implementation is currently only used by the
experimental HTTP worker for the ng media backend. Which will also
stop using it as soon as we have time to refactor it.
There is no need to bloat the production code by keeping it there.
