+++ tests/zyppng/media/CMakeLists.txt
@@ -1,7 +1,7 @@
 IF( NOT DISABLE_MEDIABACKEND_TESTS)
   ADD_TESTS(
     NetworkRequestDispatcher
-    EvDownloader
+    #EvDownloader
     Provider
   )
   target_link_libraries( Provider_test PUBLIC tvm-protocol-obj )
