+++ tools/zypp-media-http/downloader/downloadspec.h
