+++ tools/zypp-media-http/downloader/private/downloader_p.h
