+++ tools/zypp-media-http/downloader/private/downloaderstates/basicdownloader_p.h
