+++ tools/zypp-media-http/downloader/private/downloaderstates/detectmeta_p.h
