+++ tools/zypp-media-http/downloader/private/downloaderstates/initial_p.h
