+++ tools/zypp-media-http/downloader/private/downloaderstates/metalinkinfo_p.h
