+++ tools/zypp-media-http/downloader/private/downloaderstates/mirrorhandling_p.h
