+++ tools/zypp-media-http/downloader/private/downloaderstates/normal_p.h
