+++ tools/zypp-media-http/downloader/private/downloaderstates/rangedownloader_p.h
