+++ tools/zypp-media-http/downloader/private/downloaderstates/zck_p.cc
@@ -9,7 +9,7 @@
 
 #if ENABLE_ZCHUNK_COMPRESSION
 
-#include <zypp-curl/ng/network/private/downloader_p.h>
+#include <downloader/private/downloader_p.h>
 #include <zypp-curl/ng/network/private/mediadebug_p.h>
 #include <zypp-curl/ng/network/private/networkrequesterror_p.h>
 #include <zypp-core/AutoDispose.h>
