+++ tools/zypp-media-http/downloader/private/downloaderstates/zck_p.h
