+++ tools/zypp-media-http/downloader/private/mirrorcontrol_p.h
