+++ zypp-curl/CMakeLists.txt
@@ -64,9 +64,6 @@ INSTALL(  FILES ${zypp_curl_proxyinfo_HEADERS} DESTINATION ${INCLUDE_INSTALL_DIR
 
 SET( zypp_curl_ng_network_SRCS
   ng/network/curlmultiparthandler.cc
-  ng/network/downloader.cc
-  ng/network/downloadspec.cc
-  ng/network/mirrorcontrol.cc
   ng/network/networkrequestdispatcher.cc
   ng/network/networkrequesterror.cc
   ng/network/request.cc
@@ -76,10 +73,6 @@ SET( zypp_curl_ng_network_HEADERS
   ng/network/curlmultiparthandler.h
   ng/network/AuthData
   ng/network/authdata.h
-  ng/network/Downloader
-  ng/network/downloader.h
-  ng/network/DownloadSpec
-  ng/network/downloadspec.h
   ng/network/NetworkRequestDispatcher
   ng/network/networkrequestdispatcher.h
   ng/network/NetworkRequestError
@@ -91,38 +84,10 @@ SET( zypp_curl_ng_network_HEADERS
 )
 
 SET( zypp_curl_ng_network_private_HEADERS
-  ng/network/private/downloader_p.h
   ng/network/private/mediadebug_p.h
-  ng/network/private/mirrorcontrol_p.h
   ng/network/private/networkrequestdispatcher_p.h
   ng/network/private/networkrequesterror_p.h
   ng/network/private/request_p.h
-  ng/network/private/downloaderstates/base_p.h
-  ng/network/private/downloaderstates/basicdownloader_p.h
-  ng/network/private/downloaderstates/detectmeta_p.h
-  ng/network/private/downloaderstates/final_p.h
-  ng/network/private/downloaderstates/initial_p.h
-  ng/network/private/downloaderstates/metalinkinfo_p.h
-  ng/network/private/downloaderstates/metalink_p.h
-  ng/network/private/downloaderstates/mirrorhandling_p.h
-  ng/network/private/downloaderstates/normal_p.h
-  ng/network/private/downloaderstates/preparemulti_p.h
-  ng/network/private/downloaderstates/rangedownloader_p.h
-  ng/network/private/downloaderstates/zck_p.h
-)
-
-SET( zypp_curl_ng_network_private_SRCS
-  ng/network/private/downloaderstates/basicdownloader_p.cc
-  ng/network/private/downloaderstates/detectmeta_p.cc
-  ng/network/private/downloaderstates/final_p.cc
-  ng/network/private/downloaderstates/initial_p.cc
-  ng/network/private/downloaderstates/metalinkinfo_p.cc
-  ng/network/private/downloaderstates/metalink_p.cc
-  ng/network/private/downloaderstates/mirrorhandling_p.cc
-  ng/network/private/downloaderstates/normal_p.cc
-  ng/network/private/downloaderstates/preparemulti_p.cc
-  ng/network/private/downloaderstates/rangedownloader_p.cc
-  ng/network/private/downloaderstates/zck_p.cc
 )
 
 # ng files are not installed yet
