--- zypp-curl/ng/network/DownloadSpec
@@ -1,1 +1,0 @@
-#include "downloadspec.h"
