--- zypp-curl/ng/network/Downloader
@@ -1,1 +1,0 @@
-#include "downloader.h"
