--- zypp-curl/ng/network/MirrorControl
@@ -1,1 +1,0 @@
-#include "mirrorcontrol.h"
